requirements.txt
streamlit
streamlit-drawable-canvas
scikit-learn
joblib
numpy
Pillow